refactor script parser
